Diagonal Motion!
Basic Collision tests in the TestDiagMotionDriver
I did ever directing including with x and/or y = 0. They all work!
So now what needs to be done is to loop this process, while iterating a
copy of the Entity through the motion (see Phantom motion in DiagMotion
class). The next "step" should be done on the phantom, and we might not
necessarily want the object to iterate with every test...

Anyways, I did it :D