Node Overhaul & Inversor
Ability to now Inverse HashTable
Node new function CopySelf, which allows the node to copy itself in a
variety of different fashions. This caused a large refactoring overhaul
in a few places, such as Hashtable. Allow for Increasie efficency of
algo.

Uploaded file called "Notes" It mentions a few tests needed to be run,
since the new code has not been fully tested, it seems to run in all
tested cases so far.